+++ src/common/functions.ts
@@ -20,30 +20,31 @@ export const scrollIntoViewTowardsCenter = (elem: HTMLElement) => {
 
 export const interactionSelectors = 'input, select, textarea, button, object, a, area[href], [tabindex]'
 
-export const flipFocus = function (event: React.KeyboardEvent) {
+export const flipFocus = function (event: KeyboardEvent) {
   if (event.altKey && event.key === 'Tab') {
     event.stopPropagation()
 
     const contentArea = document.querySelector('section.screen > article')
     const navArea = document.querySelector('section.screen > nav')
-    const contentSelector = contentArea?.querySelector(interactionSelectors) as HTMLElement
-    const navSelectors = navArea?.querySelectorAll(interactionSelectors) as NodeListOf<HTMLElement>
-
     let focusable: HTMLElement | undefined
-    if (contentSelector && navArea?.contains(document.activeElement)) {
-      focusable = contentSelector
-    } else if (navSelectors.length && contentArea?.contains(document.activeElement)) {
-      focusable = navSelectors[navSelectors.length - 1]
-    } else if (!focusable) {
+
+    if (navArea?.contains(document.activeElement)) {
+      focusable = contentArea?.querySelector(interactionSelectors) as HTMLElement
+    } else if (contentArea?.contains(document.activeElement)) {
+      focusable = navArea?.querySelector(interactionSelectors) as HTMLElement
+    }
+
+    if (!focusable) {
       const all = document.querySelectorAll(interactionSelectors)
       if (all.length) {
-        if (all[all.length -1] === document.activeElement) {
-          focusable = all[0] as HTMLElement
-        } else {
+        if (all[0] === document.activeElement) {
           focusable = all[all.length -1] as HTMLElement
+        } else {
+          focusable = all[0] as HTMLElement
         }
       }
     }
-    focusable?.focus()
+
+    if (focusable) focusable.focus()
   }
 }